rename utils, remove import
